Organizes the source code. Nothing has been changed.
