+++ src/Elders.VSE-FormatDocumentOnSave/Configurations/EditorConfigConfiguration.cs
@@ -1,0 +1,34 @@
+﻿using EditorConfig.Core;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Elders.VSE_FormatDocumentOnSave.Configurations
+{
+    public class EditorConfigConfiguration : IConfiguration
+    {
+        string allowed = ".*";
+        string denied = "";
+        string command = "Edit.FormatDocument";
+
+        public EditorConfigConfiguration(string formatConfigFile)
+        {
+            var parser = new EditorConfig.Core.EditorConfigParser(formatConfigFile);
+            FileConfiguration configFile = parser.Parse(formatConfigFile).First();
+
+            if (configFile.Properties.ContainsKey("allowed_extensions"))
+                configFile.Properties.TryGetValue("allowed_extensions", out allowed);
+
+            if (configFile.Properties.ContainsKey("denied_extensions"))
+                configFile.Properties.TryGetValue("denied_extensions", out denied);
+
+            if (configFile.Properties.ContainsKey("command"))
+                configFile.Properties.TryGetValue("command", out command);
+        }
+
+        IEnumerable<string> IConfiguration.Allowed => allowed.Split(' ');
+
+        IEnumerable<string> IConfiguration.Denied => denied.Split(' ');
+
+        public string Command => command;
+    }
+}
