+++ src/Elders.VSE-FormatDocumentOnSave/FormatDocumentOnSavePackage.cs
@@ -8,6 +8,8 @@ using System.IO;
 using System.Linq;
 using EditorConfig.Core;
 using System;
+using Microsoft.VisualStudio;
+using Elders.VSE_FormatDocumentOnSave.Configurations;
 
 namespace Elders.VSE_FormatDocumentOnSave
 {
@@ -94,148 +96,5 @@ namespace Elders.VSE_FormatDocumentOnSave
         }
     }
 
-    public class FormatDocumentConfiguration : IConfiguration
-    {
-        private readonly IConfiguration configuration;
 
-        public FormatDocumentConfiguration(Document doc, IConfiguration defaultCfg)
-        {
-            configuration = defaultCfg;
-
-            FileInfo cfgFile = new FileInfo(Path.Combine(doc.Path, ".formatconfig"));
-            var dir = new DirectoryInfo(doc.Path);
-            while (dir.Parent != null)
-            {
-                if (cfgFile.Exists) break;
-
-                var configs = dir.GetFiles(".formatconfig");
-                if (configs.Length > 0)
-                {
-                    cfgFile = configs[0];
-                    break;
-                }
-                else
-                {
-                    dir = dir.Parent;
-                }
-            }
-
-            if (cfgFile.Exists)
-                configuration = new EditorConfigConfiguration(cfgFile.FullName);
-        }
-
-        public IEnumerable<string> Allowed => configuration.Allowed;
-
-        public IEnumerable<string> Denied => configuration.Denied;
-
-        public string Command => configuration.Command;
-    }
-
-    public interface IConfiguration
-    {
-        /// <summary>
-        /// Allowed extensions. For example: .cs .html .cshtml .vb
-        /// </summary>
-        IEnumerable<string> Allowed { get; }
-
-        /// <summary>
-        /// Denied filed extentions. For example: .cs .html .cshtml .vb
-        /// </summary>
-        IEnumerable<string> Denied { get; }
-
-        /// <summary>
-        /// The Visual Studio command to execute. Defaults to format document (Edit.FormatDocument)
-        /// </summary>
-        string Command { get; }
-    }
-
-    public class EditorConfigConfiguration : IConfiguration
-    {
-        string allowed = ".*";
-        string denied = "";
-        string command = "Edit.FormatDocument";
-
-        public EditorConfigConfiguration(string formatConfigFile)
-        {
-            var parser = new EditorConfig.Core.EditorConfigParser(formatConfigFile);
-            FileConfiguration configFile = parser.Parse(formatConfigFile).First();
-
-            if (configFile.Properties.ContainsKey("allowed_extensions"))
-                configFile.Properties.TryGetValue("allowed_extensions", out allowed);
-
-            if (configFile.Properties.ContainsKey("denied_extensions"))
-                configFile.Properties.TryGetValue("denied_extensions", out denied);
-
-            if (configFile.Properties.ContainsKey("command"))
-                configFile.Properties.TryGetValue("command", out command);
-        }
-
-        IEnumerable<string> IConfiguration.Allowed
-        {
-            get
-            {
-                return allowed.Split(' ');
-            }
-        }
-
-        IEnumerable<string> IConfiguration.Denied
-        {
-            get
-            {
-                return denied.Split(' ');
-            }
-        }
-
-        public string Command => command;
-    }
-
-    public class VisualStudioConfiguration : DialogPage, IConfiguration
-    {
-        string allowed = ".*";
-        string denied = "";
-        string command = "Edit.FormatDocument";
-
-        [Category("Format Document On Save")]
-        [DisplayName("Allowed extensions")]
-        [Description("Space separated list. For example: [.cs .html .cshtml .vb] Overrides extensions listed in Denied section. When you use [.*] the extensions listed in Denied section will be ignored. Empty value respects all extensions listed in Denied section.")]
-        public string Allowed
-        {
-            get { return allowed; }
-            set { allowed = value; }
-        }
-
-        [Category("Format Document On Save")]
-        [DisplayName("Denied extensions")]
-        [Description("Space separated list. For example: [.cs .html .cshtml .vb]")]
-        public string Denied
-        {
-            get { return denied; }
-            set { denied = value; }
-        }
-
-        [Category("Format Document On Save")]
-        [DisplayName("Command")]
-        [Description("The Visual Studio command to execute. Defaults to VS command [Edit.FormatDocument]")]
-        public string Command
-        {
-            get { return command; }
-            set { command = value; }
-        }
-
-        IEnumerable<string> IConfiguration.Allowed
-        {
-            get
-            {
-                return Allowed.Split(' ');
-            }
-        }
-
-        IEnumerable<string> IConfiguration.Denied
-        {
-            get
-            {
-                return Denied.Split(' ');
-            }
-        }
-    }
 }
